requirements/common.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
Django==1.11.7
djehouty==0.1
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Testing==0.6.2
funcsigs==1.0.2
idna==2.6
itsdangerous==0.24
html2text==2018.1.9
Jinja2==2.10
jsonfield==2.0.2
logutils==0.3.5
mail-parser==3.3.2
MarkupSafe==1.0
mock==2.0.0
netaddr==0.7.19
psycopg2==2.7.3.2
pyaml==17.10.0
pyjwt==1.6.4
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
rq==0.6.0
rq-dashboard==0.3.11
rq-scheduler==0.7.0
simplejson==3.12.0
six==1.11.0
urllib3==1.22
uWSGI==2.0.15
voluptuous==0.10.5
Werkzeug==0.12.2
